+++ src/components/Common/InputBox.stories.tsx
