+++ src/components/Template/PageTemplate.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Header from '../Header';
+import Header from '../Common/Header';
 
 interface PageTemplateProps {
   children: React.ReactNode;
